Add model-namespace
