+++ eng/packages/http-client-csharp/emitter/src/index.ts
@@ -1,4 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-export { $lib, $onEmit, AzureEmitterOptions, AzureEmitterOptionsSchema } from "./emitter.js";
+export { $onEmit } from "./emitter.js";
+export { AzureEmitterOptions, AzureEmitterOptionsSchema } from "./options.js";
+export { $lib } from "./lib/lib.js";
